move reward carousel logic into its own viewmodel/ui state
