+++ app/src/main/java/com/kickstarter/viewmodels/projectpage/CheckoutFlowViewModel.kt
@@ -8,17 +8,10 @@ import com.kickstarter.libs.Environment
 import com.kickstarter.libs.rx.transformers.Transformers
 import com.kickstarter.libs.utils.RewardUtils
 import com.kickstarter.libs.utils.extensions.addToDisposable
-import com.kickstarter.libs.utils.extensions.isBacked
 import com.kickstarter.libs.utils.extensions.isNotNull
-import com.kickstarter.mock.factories.RewardFactory
-import com.kickstarter.models.Backing
 import com.kickstarter.models.Location
-import com.kickstarter.models.Project
 import com.kickstarter.models.Reward
 import com.kickstarter.models.ShippingRule
-import com.kickstarter.ui.data.PledgeData
-import com.kickstarter.ui.data.PledgeFlowContext
-import com.kickstarter.ui.data.PledgeReason
 import com.kickstarter.ui.data.ProjectData
 import io.reactivex.Observable
 import io.reactivex.disposables.CompositeDisposable
@@ -29,18 +22,10 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
-import java.util.Locale
 
 data class FlowUIState(
     val currentPage: Int = 0,
     val expanded: Boolean = false
-)
-
-data class RewardSelectionUIState(
-    val rewardList: List<Reward> = listOf(),
-    val initialRewardIndex: Int = 0,
-    val project: ProjectData = ProjectData.builder().build(),
-    val showAlertDialog: Boolean = false
 )
 
 class CheckoutFlowViewModel(val environment: Environment) : ViewModel() {
@@ -56,20 +41,8 @@ class CheckoutFlowViewModel(val environment: Environment) : ViewModel() {
     val currentUserReward = PublishSubject.create<Reward>()
     val projectData = PublishSubject.create<ProjectData>()
 
-    lateinit var currentProjectData: ProjectData
-    var previousUserBacking: Backing? = null
-    var previouslyBackedReward: Reward? = null
-    lateinit var newUserReward: Reward
-
-    private val mutableRewardSelectionUIState = MutableStateFlow(RewardSelectionUIState())
-    val rewardSelectionUIState: StateFlow<RewardSelectionUIState>
-        get() = mutableRewardSelectionUIState
-            .asStateFlow()
-            .stateIn(
-                scope = viewModelScope,
-                started = SharingStarted.WhileSubscribed(),
-                initialValue = RewardSelectionUIState()
-            )
+    private lateinit var currentProjectData: ProjectData
+    private lateinit var newUserReward: Reward
 
     private val mutableFlowUIState = MutableStateFlow(FlowUIState())
     val flowUIState: StateFlow<FlowUIState>
@@ -100,20 +73,15 @@ class CheckoutFlowViewModel(val environment: Environment) : ViewModel() {
             }.addToDisposable(disposables)
     }
 
+    fun changePage(requestedFlowState: FlowUIState) {
+        viewModelScope.launch {
+            mutableFlowUIState.emit(requestedFlowState)
+        }
+    }
+
     fun provideProjectData(projectData: ProjectData) {
         currentProjectData = projectData
-        previousUserBacking = projectData.backing()
-        previouslyBackedReward = getReward(previousUserBacking)
-        val indexOfBackedReward = indexOfBackedReward(project = projectData.project())
         viewModelScope.launch {
-            mutableRewardSelectionUIState.emit(
-                RewardSelectionUIState(
-                    rewardList = projectData.project().rewards() ?: listOf(),
-                    initialRewardIndex = indexOfBackedReward,
-                    project = projectData
-                )
-            )
-
             projectData.project().rewards()?.let { rewards ->
                 apolloClient.getShippingRules(
                     reward = rewards.first { theOne ->
@@ -150,78 +118,7 @@ class CheckoutFlowViewModel(val environment: Environment) : ViewModel() {
     fun userRewardSelection(reward: Reward) {
         viewModelScope.launch {
             currentUserReward.onNext(reward)
-            val pledgeDataAndReason = pledgeDataAndPledgeReason(currentProjectData, reward)
-            newUserReward = pledgeDataAndReason.first.reward()
-
-            when (pledgeDataAndReason.second) {
-                PledgeReason.UPDATE_REWARD -> {
-                    if (previouslyBackedReward?.hasAddons() == true && !newUserReward.hasAddons())
-                    // Show warning to user
-                        mutableRewardSelectionUIState.emit(
-                            RewardSelectionUIState(
-                                rewardList = currentProjectData.project().rewards() ?: listOf(),
-                                project = currentProjectData,
-                                showAlertDialog = true
-                            )
-                        )
-
-                    if (previouslyBackedReward?.hasAddons() == false && !newUserReward.hasAddons())
-                    // Go to confirm page
-                        mutableFlowUIState.emit(FlowUIState(currentPage = 2, expanded = true))
-
-                    if (previouslyBackedReward?.hasAddons() == true && newUserReward.hasAddons()) {
-                        if (differentShippingTypes(previouslyBackedReward, newUserReward))
-                        // Show warning to user
-                            mutableRewardSelectionUIState.emit(
-                                RewardSelectionUIState(
-                                    rewardList = currentProjectData.project().rewards() ?: listOf(),
-                                    project = currentProjectData,
-                                    showAlertDialog = true
-                                )
-                            )
-                        // Go to add-ons
-                        else mutableFlowUIState.emit(FlowUIState(currentPage = 1, expanded = true))
-                    }
-
-                    if (previouslyBackedReward?.hasAddons() == false && newUserReward.hasAddons()) {
-                        // Go to add-ons
-                        mutableFlowUIState.emit(FlowUIState(currentPage = 1, expanded = true))
-                    }
-                }
-
-                PledgeReason.PLEDGE -> {
-                    if (newUserReward.hasAddons())
-                    // Show add-ons
-                        mutableFlowUIState.emit(FlowUIState(currentPage = 1, expanded = true))
-                    else
-                    // Show confirm page
-                        mutableFlowUIState.emit(FlowUIState(currentPage = 2, expanded = true))
-                }
-
-                else -> {
-                }
-            }
-        }
-    }
-
-    fun onRewardCarouselAlertClicked(wasPositive: Boolean) {
-        viewModelScope.launch {
-            mutableRewardSelectionUIState.emit(
-                RewardSelectionUIState(
-                    rewardList = currentProjectData.project().rewards() ?: listOf(),
-                    project = currentProjectData,
-                    showAlertDialog = false
-                )
-            )
-            if (wasPositive) {
-                if (newUserReward.hasAddons()) {
-                    // Go to add-ons
-                    mutableFlowUIState.emit(FlowUIState(currentPage = 1, expanded = true))
-                } else {
-                    // Show confirm page
-                    mutableFlowUIState.emit(FlowUIState(currentPage = 2, expanded = true))
-                }
-            }
+            newUserReward = reward
         }
     }
 
@@ -230,46 +127,6 @@ class CheckoutFlowViewModel(val environment: Environment) : ViewModel() {
             // Show confirm page
             mutableFlowUIState.emit(FlowUIState(currentPage = 2, expanded = true))
         }
-    }
-
-    private fun pledgeDataAndPledgeReason(
-        projectData: ProjectData,
-        reward: Reward
-    ): Pair<PledgeData, PledgeReason> {
-        val pledgeReason =
-            if (projectData.project().isBacking()) PledgeReason.UPDATE_REWARD
-            else PledgeReason.PLEDGE
-        val pledgeData =
-            PledgeData.with(PledgeFlowContext.forPledgeReason(pledgeReason), projectData, reward)
-        return Pair(pledgeData, pledgeReason)
-    }
-
-    private fun differentShippingTypes(newRW: Reward?, backedRW: Reward): Boolean {
-        return if (newRW == null) false
-        else if (newRW.id() == backedRW.id()) false
-        else {
-            (newRW.shippingType()?.lowercase(Locale.getDefault()) ?: "") != (backedRW.shippingType()?.lowercase(Locale.getDefault()) ?: "")
-        }
-    }
-
-    private fun getReward(backingObj: Backing?): Reward? {
-        backingObj?.let { backing ->
-            return backing.reward()?.let { reward ->
-                if (backing.addOns().isNullOrEmpty()) reward
-                else reward.toBuilder().hasAddons(true).build()
-            } ?: RewardFactory.noReward()
-        } ?: return null
-    }
-
-    private fun indexOfBackedReward(project: Project): Int {
-        project.rewards()?.run {
-            for ((index, reward) in withIndex()) {
-                if (project.backing()?.isBacked(reward) == true) {
-                    return index
-                }
-            }
-        }
-        return 0
     }
 
     fun onBackPressed(currentPage: Int) {
